requirements.txt
requests
discord.py
aiohttp
websockets
chardet
asyncio